Map known beacon IDs to their names in the tree diff
