+++ frontend/components/merkle-tree-elements/index.tsx
@@ -1,15 +1,15 @@
-import { useEffect, useRef, useState } from 'react';
+import { useRef } from 'react';
 import some from 'lodash/some';
-import { Diff2HtmlUI } from 'diff2html/lib/ui/js/diff2html-ui';
 import { AlertTriangleIcon, InfoIcon, ShieldCheckIcon, ShieldEllipsisIcon } from 'lucide-react';
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '~/components/ui/table';
 import { Tooltip, TooltipContent, TooltipTrigger } from '~/components/ui/tooltip';
 import { Button } from '~/components/ui/button';
-import { Toggle } from '~/components/ui/toggle';
 import { cn } from '~/lib/utils';
 import { useWeb3Data } from '~/contexts/web3-data-context';
 import addressBook from '../../../data/address-book.json';
 import 'diff2html/bundles/css/diff2html.min.css';
+
+export { TreeDiff } from './tree-diff';
 
 interface TreeStatusBadgeProps {
   signatures: Record<string, string>;
@@ -142,7 +142,6 @@ function SignerInfo(props: SignerInfoProps) {
         </Button>
       </TooltipTrigger>
       <TooltipContent
-        className="border-0 bg-gray-600 text-xs text-gray-200"
         onPointerDownOutside={(ev) => {
           // We don't want to close the tooltip when the trigger is clicked
           if (triggerRef.current!.contains(ev.target as Node)) {
@@ -153,82 +152,6 @@ function SignerInfo(props: SignerInfoProps) {
         {props.address}
       </TooltipContent>
     </Tooltip>
-  );
-}
-
-interface TreeDiffProps {
-  diffResult: null | { diff: string; status: 'success' } | { status: 'error' };
-}
-
-type DiffMode = null | 'unified' | 'split';
-
-export function TreeDiff(props: TreeDiffProps) {
-  const { diffResult } = props;
-
-  const [mode, setMode] = useState<DiffMode>(null);
-  useEffect(() => {
-    const storedMode = window.localStorage.getItem('diff-mode');
-    setMode((storedMode || 'split') as DiffMode);
-  }, []);
-
-  useEffect(() => {
-    if (!mode) return;
-
-    if (diffResult?.status === 'success' && diffResult.diff) {
-      const element = document.getElementById('tree-diff-container')!;
-      const ui = new Diff2HtmlUI(element, diffResult.diff, {
-        drawFileList: false,
-        fileContentToggle: false,
-        synchronisedScroll: true,
-        outputFormat: mode === 'unified' ? 'line-by-line' : 'side-by-side',
-        rawTemplates: { 'tag-file-renamed': '' },
-      });
-      ui.draw();
-      ui.highlightCode();
-    }
-  }, [diffResult, mode]);
-
-  const previousFile = <span className="font-semibold">previous-hash.json</span>;
-  const currentFile = <span className="font-semibold">current-hash.json</span>;
-
-  return (
-    <div>
-      {diffResult == null ? (
-        <p className="my-4 text-sm text-gray-500">
-          There is no {previousFile} file to compare the {currentFile} file with.
-        </p>
-      ) : diffResult.status === 'error' ? (
-        <p className="my-4 inline-block rounded bg-red-100 px-4 py-3 text-sm text-red-700">
-          Something went wrong comparing {previousFile} and {currentFile}.
-        </p>
-      ) : diffResult.diff === '' ? (
-        <p className="my-4 text-sm text-gray-500">
-          The contents of {previousFile} and {currentFile} are identical.
-        </p>
-      ) : (
-        <>
-          <div className="flex items-center justify-between gap-4">
-            <p className="my-4 text-sm text-gray-500">
-              Shows the difference between the {previousFile} and the {currentFile} files.
-            </p>
-            <Toggle
-              variant="outline"
-              size="sm"
-              onPressedChange={(unified) => {
-                const newMode = unified ? 'unified' : 'split';
-                setMode(newMode);
-                window.localStorage.setItem('diff-mode', newMode);
-              }}
-              pressed={mode === 'unified'}
-              aria-label="Toggle unified"
-            >
-              Unified
-            </Toggle>
-          </div>
-          <div id="tree-diff-container" className="w-full overflow-x-auto" />
-        </>
-      )}
-    </div>
   );
 }
